+++ src/Modules/Vocabulary/Application/Helpers/StatusHelper.php
@@ -20,6 +20,7 @@ namespace Lwt\Modules\Vocabulary\Application\Helpers;
 
 use Lwt\Shared\UI\Helpers\IconHelper;
 use Lwt\Modules\Vocabulary\Application\Services\TermStatusService;
+use Lwt\Modules\Vocabulary\Domain\ValueObject\TermStatus;
 
 /**
  * Helper class for word status-related display logic.
@@ -111,7 +112,7 @@ class StatusHelper
             return '';
         }
 
-        $allStatuses = [1, 2, 3, 4, 5, 98, 99];
+        $allStatuses = TermStatus::values();
         $includedStatuses = self::getIncludedStatuses($status);
 
         // Build :not() selectors for statuses NOT in the filter